rename step and fix some tests referencing the old name
