+++ service/src/main/java/bio/terra/profile/service/profile/flight/create/CreateProfileFlight.java
@@ -9,7 +9,7 @@ import bio.terra.profile.service.crl.GcpCrlService;
 import bio.terra.profile.service.iam.SamService;
 import bio.terra.profile.service.job.JobMapKeys;
 import bio.terra.profile.service.policy.TpsApiDispatch;
-import bio.terra.profile.service.profile.flight.common.VerifyAccountUserAccessStep;
+import bio.terra.profile.service.profile.flight.common.VerifyUserBillingAccountAccessStep;
 import bio.terra.profile.service.profile.model.BillingProfile;
 import bio.terra.profile.service.profile.model.ProfileDescription;
 import bio.terra.stairway.Flight;
@@ -40,7 +40,7 @@ public class CreateProfileFlight extends Flight {
     addStep(new CreateProfileStep(profileDao, billingProfile, user));
     switch (billingProfile.cloudPlatform()) {
       case GCP:
-        addStep(new VerifyAccountUserAccessStep(crlService, billingProfile, user));
+        addStep(new VerifyUserBillingAccountAccessStep(crlService, billingProfile, user));
         break;
       case AZURE:
         addStep(
